POC: Build a trampoline function
Add a new function foo that calls printf with some arguments. Now the plugin
replaces the address of printf in the call in foo() with the address of the
generated function but doesn't modify the call arguments or return value. The
generated function itself contains only a call to the replaced function address
with no arguments nor a return value.